unify test_funnel and test_funnel_udf
